+++ .gitignore
@@ -28,3 +28,5 @@ CLAUDE.md
 
 .cursorignore
 .cursorrules
+
+.env
