+++ src/components/talkingkit/layout/Step1Layout.tsx
@@ -1,9 +1,9 @@
-import { ReactNode } from 'react';
-import LeftArrowIcon from '@/assets/svgs/talkingkit/leftarrow.svg';
-import AnimatedContainer from '@/components/talkingkit/AnimatedContainer';
+import type { ReactNode } from 'react';
+import LeftArrowIcon from '@/assets/svgs/talkingkit/common/leftarrow.svg';
+import AnimatedContainer from '@/components/talkingkit/common/AnimatedContainer';
 import clsx from 'clsx';
 
-interface Step2LayoutProps {
+interface Step1LayoutProps {
   headerTitle: string;
   title: string;
   children: ReactNode;
@@ -15,7 +15,7 @@ interface Step2LayoutProps {
   disableAnimation?: boolean;
 }
 
-const Step2Layout = ({
+const Step1Layout = ({
   headerTitle,
   title,
   children,
@@ -25,7 +25,7 @@ const Step2Layout = ({
   onButtonClick,
   onBackClick,
   disableAnimation = false,
-}: Step2LayoutProps) => {
+}: Step1LayoutProps) => {
   return (
     <div className="bg-background-primary relative flex h-full flex-col">
       {/* 상단 바 */}
@@ -46,22 +46,27 @@ const Step2Layout = ({
         </div>
       </div>
 
-      {/* 상단 진행바 (2단계 활성) */}
+      {/* 상단 진행바 (1단계 활성) */}
       <AnimatedContainer variant="fadeInUpSmall" delay={0} className="px-4 py-3" disabled={disableAnimation}>
         <div className="flex gap-2">
-          <div className="h-1 flex-1 rounded-full bg-gray-200" />
           <div className="bg-blue-1 h-1 flex-1 rounded-full" />
+          <div className="h-1 flex-1 rounded-full bg-gray-200" />
         </div>
       </AnimatedContainer>
 
       {/* 본문 */}
       <div className="relative flex flex-1 flex-col items-center px-4 py-4 pb-12">
         {/* 단계 정보와 흰색 박스 */}
-        <div className={clsx('flex w-full flex-col', showAction ? 'gap-[72px]' : 'gap-0')}>
+        <div className={clsx('flex w-full flex-col', showAction ? 'gap-10' : 'gap-0')}>
           {/* 단계 정보 */}
           <div className="flex w-full flex-col gap-2">
-            <AnimatedContainer variant="fadeInUp" delay={0.1} className="mb-2 w-full text-left" disabled={disableAnimation}>
-              <p className="text-detail-01 text-gray-60">2단계</p>
+            <AnimatedContainer
+              variant="fadeInUp"
+              delay={0.1}
+              className="mb-2 w-full text-left"
+              disabled={disableAnimation}
+            >
+              <p className="text-detail-01 text-gray-60">1단계</p>
               <h2 className="text-heading-02-semibold text-gray-100">{title}</h2>
             </AnimatedContainer>
 
@@ -78,7 +83,12 @@ const Step2Layout = ({
 
           {/* 안내 텍스트와 버튼 (조건부) */}
           {showAction && (
-            <AnimatedContainer variant="fadeIn" delay={0.25} className="flex w-full flex-col items-center gap-4" disabled={disableAnimation}>
+            <AnimatedContainer
+              variant="fadeIn"
+              delay={0.25}
+              className="flex w-full flex-col items-center gap-4"
+              disabled={disableAnimation}
+            >
               {/* 안내 텍스트 */}
               {guideText && <p className="text-body-02-regular text-gray-60 text-center">{guideText}</p>}
 
@@ -99,4 +109,4 @@ const Step2Layout = ({
   );
 };
 
-export default Step2Layout;
+export default Step1Layout;
