+++ src/components/talkingkit/loudSound/DecibelBar.tsx
