+++ src/components/talkingkit/progressBar/ProgressBar.tsx
