+++ src/components/talkingkit/progressBar/TimerProgressBar.tsx
