+++ src/pages/talkingkit/TalkingKit.tsx
@@ -1,5 +1,5 @@
-import LeftArrowIcon from '@/assets/svgs/talkingkit/leftarrow.svg';
-import KitCard from '@/components/talkingkit/KitCard';
+import LeftArrowIcon from '@/assets/svgs/talkingkit/common/leftarrow.svg';
+import KitCard from '@/components/talkingkit/common/KitCard';
 import { kitsData } from '@/mock/talkingkit/kitsData';
 
 const TalkingKit = () => {
